Bug 509194 - Invalid menu extension contributions
This change replaces action menu definitions with command definitions,
for the following actions:

* org.eclipse.dltk.ui.actions.Move
* org.eclipse.dltk.ui.actions.Rename

This is done to avoid menu extension errors on perspective
customization, due to problematic code in CustomizePerspectiveDialog. In
particular the customize perspective dialog will populate main menu
submenus only after going over actions; this causes the menu paths of
the actions above to be detected as invalid, despite actually being
valid and functional.

Change-Id: Id6ae0b94d6e0bd7329d2191a7527b684e5ed56e2
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>